created separate notifications api file so that accounts api isn't so bloated
