+++ api/api.py
@@ -11,6 +11,7 @@ from fastapi.encoders import jsonable_encoder
 import auth
 import account_api
 import initiatives_api
+import notifications_api
 from settings import Config, Session, get_db
 import logging
 from uuid import uuid4, UUID
@@ -41,6 +42,10 @@ app.include_router(
 )
 app.include_router(
     initiatives_api.router,
+    prefix='/api'
+)
+app.include_router(
+    notifications_api.router,
     prefix='/api'
 )
